+++ Headers/AppKit/NSTextContainer.h
@@ -97,7 +97,7 @@ typedef NSUInteger NSLineMovementDirection;
   id _textView;
 
   NSRect _containerRect;
-  float _lineFragmentPadding;
+  CGFloat _lineFragmentPadding;
 
   BOOL _observingFrameChanges;
   BOOL _widthTracksTextView;
@@ -260,8 +260,8 @@ Line fragment padding<br />
 The line fragment padding is an amount of space left empty at each end of
 a line fragment rectangle by the standard typesetter. The default is 0.0.
 */
-- (void) setLineFragmentPadding: (float)aFloat;
-- (float) lineFragmentPadding;
+- (void) setLineFragmentPadding: (CGFloat)aFloat;
+- (CGFloat) lineFragmentPadding;
 
 @end
